pyproject.toml
[tool.poetry]
name = "drf-clientes-api"
version = "0.1.0"
description = ""
authors = ["sabrina"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
